app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
opencv-python==4.8.1.78
passlib==1.7.4
Pillow==10.0.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyodbc==4.0.39
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
segno==1.5.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
websockets==11.0.3
mysql-connector-python==8.0.15
PyJWT==2.8.0